initApplicationStatsRouter for applications tools
Remove the need for getApplicationMappings and pass config to a router
in the same way as we do for the forms themselves.
